Handle double promotion (#407)
* Replace M_PI which is a double value with F_PI. Remove unwanted double promotion
* Remove unwanted double promotion inside KOS
* Addressed discord input
* More explicit float types
* Fixed mistake removal of f
* Avoid it being inlined into potentially another FPU context
* Hande printf related double premotions
* Clarify floating values in fmath_base

+++ examples/dreamcast/kgl/basic/gl/gltest.c
@@ -236,7 +236,7 @@ int main(int argc, char **argv) {
 
     pvr_get_stats(&stats);
     printf("VBL Count: %ld, last_time: %d, frame rate: %f fps\n",
-           stats.vbl_count, stats.frame_last_time, stats.frame_rate);
+           stats.vbl_count, stats.frame_last_time, (double)stats.frame_rate);
 
     return 0;
 }
