+++ examples/dreamcast/pvr/pvrmark_strips_direct/pvrmark_strips_direct.c
@@ -34,7 +34,7 @@ void stats(void) {
 
     pvr_get_stats(&stats);
     dbglog(DBG_DEBUG, "3D Stats: %ld frames, frame rate ~%f fps\n",
-           stats.vbl_count, stats.frame_rate);
+           stats.vbl_count, (double)stats.frame_rate);
 }
 
 
@@ -151,7 +151,7 @@ void check_switch(void) {
 
     if(now >= (begin + 5)) {
         begin = time(NULL);
-        printf("  Average Frame Rate: ~%f fps (%d pps)\n", avgfps, (int)(polycnt * avgfps));
+        printf("  Average Frame Rate: ~%f fps (%d pps)\n", (double)avgfps, (int)(polycnt * avgfps));
 
         switch(phase) {
             case PHASE_HALVE:
